Add HSTS headers to the Status App so browser use HTTPS all the time
https://en.wikipedia.org/wiki/HTTP_Strict_Transport_Security

HSTS stands for 'HTTP Strict Transport Security', and it's a way for sites
to make sure that browsers *only* connect to them over https - so if a user
enters simply `status.ophan.co.uk` the BROWSER will correct that to
https://status.ophan.co.uk/ - rather than attempting to connect to
http://status.ophan.co.uk/ (which just hangs indefinitely, our ELB's don't
listen on HTTP port 80).

This follows on from https://github.com/guardian/play-googleauth/pull/68 -
that didn't pass review because it doesn't highlight to developers what's
happening - it was an attempt to sneak HSTS headers in through the back
door!