Switch to ARM and fix Prout build-info
ARM is cheaper and faster than Intel. We're going to use the Ophan
Ubuntu Amigo images for now - only the Ophan team is using the status
app at the moment, so this feels reasonable.

Prout wasn't picking up recent PR changes - because the build info
exposed in the manifest was based on Travis - and Status App is now
built on Teamcity, rather than Travis.